Fancy arguments and other improvements
Fancy arguments via argparse. Also, prior to this change, the program expected a directory of only json/yaml files. Now, it expects a directory with any number of subdirs (which may also have subdirs, and so on) which have json/yaml files. This is easier to use for converting the entire BDFR output with multiple subreddits.